Create utils/compatibility to avoid a circular dep
