+++ tests/unit/core/utils/test_index.js
@@ -1,6 +1,5 @@
 import url from 'url';
 
-import { oneLine } from 'common-tags';
 import React from 'react';
 import config from 'config';
 import { sprintf } from 'jed';
@@ -9,7 +8,6 @@ import {
   findRenderedComponentWithType,
 } from 'react-addons-test-utils';
 import { compose } from 'redux';
-import UAParser from 'ua-parser-js';
 
 import * as api from 'core/api';
 import {
@@ -22,13 +20,6 @@ import {
   CATEGORY_COLORS,
   CLIENT_APP_ANDROID,
   CLIENT_APP_FIREFOX,
-  INCOMPATIBLE_FIREFOX_FOR_IOS,
-  INCOMPATIBLE_NO_OPENSEARCH,
-  INCOMPATIBLE_NOT_FIREFOX,
-  INCOMPATIBLE_OVER_MAX_VERSION,
-  INCOMPATIBLE_UNDER_MIN_VERSION,
-  INCOMPATIBLE_UNSUPPORTED_PLATFORM,
-  OS_MAC,
   validAddonTypes,
 } from 'core/constants';
 import {
@@ -40,12 +31,9 @@ import {
   findAddon,
   getCategoryColor,
   getClientApp,
-  getClientCompatibility,
   getClientConfig,
-  getCompatibleVersions,
   isAddonAuthor,
   isAllowedOrigin,
-  isCompatibleWithUserAgent,
   isValidClientApp,
   loadAddonIfNeeded,
   ngettext,
@@ -71,7 +59,6 @@ import {
   fakeI18n,
   unexpectedSuccess,
   userAgents,
-  userAgentsByPlatform,
 } from 'tests/unit/helpers';
 
 
@@ -321,276 +308,6 @@ describe('isAddonAuthor', () => {
 
     expect(isAddonAuthor({ addon: partialAddon, userId: null }))
       .toEqual(false);
-  });
-});
-
-describe('isCompatibleWithUserAgent', () => {
-  it('should throw if no userAgentInfo supplied', () => {
-    expect(() => {
-      isCompatibleWithUserAgent({ userAgent: null, reason: null });
-    }).toThrowError('userAgentInfo is required');
-  });
-
-  it('is incompatible with Android/webkit', () => {
-    userAgents.androidWebkit.forEach((userAgent) => {
-      expect(isCompatibleWithUserAgent({ userAgentInfo: UAParser(userAgent) }))
-        .toEqual({ compatible: false, reason: INCOMPATIBLE_NOT_FIREFOX });
-    });
-  });
-
-  it('is incompatible with Chrome Android', () => {
-    userAgents.chromeAndroid.forEach((userAgent) => {
-      expect(isCompatibleWithUserAgent({ userAgentInfo: UAParser(userAgent) }))
-        .toEqual({ compatible: false, reason: INCOMPATIBLE_NOT_FIREFOX });
-    });
-  });
-
-  it('is incompatible with Chrome desktop', () => {
-    userAgents.chrome.forEach((userAgent) => {
-      expect(isCompatibleWithUserAgent({ userAgentInfo: UAParser(userAgent) }))
-        .toEqual({ compatible: false, reason: INCOMPATIBLE_NOT_FIREFOX });
-    });
-  });
-
-  it('is compatible with Firefox desktop', () => {
-    userAgents.firefox.forEach((userAgent) => {
-      expect(isCompatibleWithUserAgent({
-        addon: createInternalAddon(fakeAddon),
-        userAgentInfo: UAParser(userAgent),
-      }))
-        .toEqual({ compatible: true, reason: null });
-    });
-  });
-
-  it('is compatible with Firefox Android', () => {
-    userAgents.firefoxAndroid.forEach((userAgent) => {
-      expect(isCompatibleWithUserAgent({
-        addon: createInternalAddon(fakeAddon),
-        userAgentInfo: UAParser(userAgent),
-      }))
-        .toEqual({ compatible: true, reason: null });
-    });
-  });
-
-  it('is compatible with Firefox OS', () => {
-    userAgents.firefoxOS.forEach((userAgent) => {
-      expect(isCompatibleWithUserAgent({
-        addon: createInternalAddon(fakeAddon),
-        userAgentInfo: UAParser(userAgent),
-      }))
-        .toEqual({ compatible: true, reason: null });
-    });
-  });
-
-  it('is incompatible with Firefox iOS', () => {
-    userAgents.firefoxIOS.forEach((userAgent) => {
-      expect(isCompatibleWithUserAgent({
-        addon: createInternalAddon(fakeAddon),
-        userAgentInfo: UAParser(userAgent),
-      }))
-        .toEqual({ compatible: false, reason: INCOMPATIBLE_FIREFOX_FOR_IOS });
-    });
-  });
-
-  it(oneLine`should use a Firefox for iOS reason code even if minVersion is
-    also not met`, () => {
-      const userAgentInfo = {
-        browser: { name: 'Firefox', version: '8.0' },
-        os: { name: 'iOS' },
-      };
-      expect(isCompatibleWithUserAgent({
-        addon: createInternalAddon(fakeAddon),
-        minVersion: '9.0',
-        userAgentInfo,
-      }))
-        .toEqual({ compatible: false, reason: INCOMPATIBLE_FIREFOX_FOR_IOS });
-    });
-
-  it('should mark Firefox without window.external as incompatible', () => {
-    const userAgentInfo = {
-      browser: { name: 'Firefox' },
-      os: { name: 'Windows' },
-    };
-    const fakeOpenSearchAddon = createInternalAddon({
-      ...fakeAddon, type: ADDON_TYPE_OPENSEARCH,
-    });
-    const fakeWindow = {};
-
-    expect(isCompatibleWithUserAgent({
-      _window: fakeWindow, addon: fakeOpenSearchAddon, userAgentInfo }))
-      .toEqual({ compatible: false, reason: INCOMPATIBLE_NO_OPENSEARCH });
-  });
-
-  it('should mark Firefox without OpenSearch support as incompatible', () => {
-    const userAgentInfo = {
-      browser: { name: 'Firefox' },
-      os: { name: 'Windows' },
-    };
-    const fakeOpenSearchAddon = createInternalAddon({
-      ...fakeAddon, type: ADDON_TYPE_OPENSEARCH,
-    });
-    const fakeWindow = { external: {} };
-
-    expect(isCompatibleWithUserAgent({
-      _window: fakeWindow, addon: fakeOpenSearchAddon, userAgentInfo }))
-      .toEqual({ compatible: false, reason: INCOMPATIBLE_NO_OPENSEARCH });
-  });
-
-  it('should mark Firefox with OpenSearch support as compatible', () => {
-    const userAgentInfo = {
-      browser: { name: 'Firefox' },
-      os: { name: 'Windows' },
-    };
-    const fakeOpenSearchAddon = createInternalAddon({
-      ...fakeAddon, type: ADDON_TYPE_OPENSEARCH,
-    });
-    const fakeWindow = { external: { AddSearchProvider: sinon.stub() } };
-
-    expect(isCompatibleWithUserAgent({
-      _window: fakeWindow, addon: fakeOpenSearchAddon, userAgentInfo }))
-      .toEqual({ compatible: true, reason: null });
-  });
-
-  it('should mark non-Firefox UAs as incompatible', () => {
-    const userAgentInfo = { browser: { name: 'Chrome' } };
-    expect(isCompatibleWithUserAgent({
-      addon: createInternalAddon(fakeAddon), userAgentInfo,
-    }))
-      .toEqual({ compatible: false, reason: INCOMPATIBLE_NOT_FIREFOX });
-  });
-
-  it('should mark Firefox 10 as incompatible with a minVersion of 10.1', () => {
-    const userAgentInfo = {
-      browser: { name: 'Firefox', version: '10.0' },
-      os: { name: 'Windows' },
-    };
-    expect(isCompatibleWithUserAgent({
-      addon: createInternalAddon(fakeAddon),
-      minVersion: '10.1',
-      userAgentInfo,
-    }))
-      .toEqual({ compatible: false, reason: INCOMPATIBLE_UNDER_MIN_VERSION });
-  });
-
-  it('should mark Firefox 24 as compatible with a maxVersion of 8', () => {
-    // https://github.com/mozilla/addons-frontend/issues/2074
-    const userAgentInfo = {
-      browser: { name: 'Firefox', version: '24.0' },
-      os: { name: 'Windows' },
-    };
-    expect(isCompatibleWithUserAgent({
-      addon: createInternalAddon({
-        ...fakeAddon,
-        current_version: {
-          ...fakeAddon.current_version,
-          is_strict_compatibility_enabled: false,
-        },
-      }),
-      maxVersion: '8',
-      userAgentInfo,
-    })).toEqual({ compatible: true, reason: null });
-  });
-
-  it('should mark Firefox as compatible when no min or max version', () => {
-    const userAgentInfo = {
-      browser: { name: 'Firefox', version: '10.0' },
-      os: { name: 'Windows' },
-    };
-    expect(isCompatibleWithUserAgent({
-      addon: createInternalAddon(fakeAddon), userAgentInfo,
-    }))
-      .toEqual({ compatible: true, reason: null });
-  });
-
-  it('should mark Firefox as compatible with maxVersion of "*"', () => {
-    // WebExtensions are marked as having a maxVersion of "*" by addons-server
-    // if their manifests don't contain explicit version information.
-    const userAgentInfo = {
-      browser: { name: 'Firefox', version: '54.0' },
-      os: { name: 'Windows' },
-    };
-    expect(isCompatibleWithUserAgent({
-      addon: createInternalAddon(fakeAddon),
-      maxVersion: '*',
-      userAgentInfo,
-    }))
-      .toEqual({ compatible: true, reason: null });
-  });
-
-  it('should log warning when minVersion is "*"', () => {
-    // Note that this should never happen as addons-server will mark a
-    // WebExtension with no minVersion as having a minVersion of "48".
-    // Still, we accept it (but it will log a warning).
-    const fakeLog = { error: sinon.stub() };
-    const userAgentInfo = {
-      browser: { name: 'Firefox', version: '54.0' },
-      os: { name: 'Windows' },
-    };
-    expect(isCompatibleWithUserAgent({
-      _log: fakeLog,
-      addon: createInternalAddon(fakeAddon),
-      minVersion: '*',
-      userAgentInfo,
-    }))
-      .toEqual({ compatible: false, reason: INCOMPATIBLE_UNDER_MIN_VERSION });
-    expect(fakeLog.error.firstCall.args[0])
-      .toContain('minVersion of "*" was passed to isCompatibleWithUserAgent()');
-  });
-
-  it('is incompatible with empty user agent values', () => {
-    const userAgentInfo = { browser: { name: '' } };
-    expect(isCompatibleWithUserAgent({
-      addon: createInternalAddon(fakeAddon), userAgentInfo,
-    }))
-      .toEqual({ compatible: false, reason: INCOMPATIBLE_NOT_FIREFOX });
-  });
-
-  it('is incompatible with non-string user agent values', () => {
-    const userAgentInfo = { browser: { name: null }, os: { name: null } };
-    expect(isCompatibleWithUserAgent({
-      addon: createInternalAddon(fakeAddon), userAgentInfo,
-    }))
-      .toEqual({ compatible: false, reason: INCOMPATIBLE_NOT_FIREFOX });
-  });
-
-  it('is incompatible if no matching platform file exists', () => {
-    const addon = createInternalAddon({
-      ...fakeAddon,
-      current_version: {
-        ...fakeAddon.current_version,
-        files: [{
-          ...fakeAddon.current_version.files[0],
-          platform: OS_MAC,
-        }],
-      },
-    });
-    const userAgentInfo =
-      UAParser(userAgentsByPlatform.windows.firefox40);
-
-    expect(isCompatibleWithUserAgent({ addon, userAgentInfo }))
-      .toEqual({
-        compatible: false,
-        reason: INCOMPATIBLE_UNSUPPORTED_PLATFORM,
-      });
-  });
-
-  it('allows non-extensions to have mismatching platform files', () => {
-    const addon = createInternalAddon({
-      ...fakeAddon,
-      current_version: {
-        ...fakeAddon.current_version,
-        files: [{
-          ...fakeAddon.current_version.files[0],
-          platform: OS_MAC,
-        }],
-      },
-      type: ADDON_TYPE_THEME,
-    });
-    const userAgentInfo =
-      UAParser(userAgentsByPlatform.windows.firefox40);
-
-    expect(isCompatibleWithUserAgent({ addon, userAgentInfo }))
-      .toMatchObject({ compatible: true });
   });
 });
 
@@ -1014,246 +731,6 @@ describe('render404IfConfigKeyIsFalse', () => {
     const props = SomeComponent.firstCall.args[0];
     expect(props.color).toEqual('orange');
     expect(props.size).toEqual('large');
-  });
-});
-
-describe('getCompatibleVersions', () => {
-  it('gets the min and max versions', () => {
-    const addon = createInternalAddon({
-      ...fakeAddon,
-      current_version: {
-        ...fakeAddon.current_version,
-        compatibility: {
-          firefox: {
-            max: '20.0.*',
-            min: '11.0.1',
-          },
-        },
-      },
-    });
-    const { maxVersion, minVersion } = getCompatibleVersions({
-      addon, clientApp: CLIENT_APP_FIREFOX });
-
-    expect(maxVersion).toEqual('20.0.*');
-    expect(minVersion).toEqual('11.0.1');
-  });
-
-  it('gets null if the clientApp does not match', () => {
-    const addon = createInternalAddon({
-      ...fakeAddon,
-      current_version: {
-        ...fakeAddon.current_version,
-        compatibility: {
-          firefox: {
-            max: '20.0.*',
-            min: '11.0.1',
-          },
-        },
-      },
-    });
-    const { maxVersion, minVersion } = getCompatibleVersions({
-      addon, clientApp: CLIENT_APP_ANDROID });
-
-    expect(maxVersion).toEqual(null);
-    expect(minVersion).toEqual(null);
-  });
-
-  it('returns null if clientApp has no compatibility', () => {
-    const addon = createInternalAddon({
-      ...fakeAddon,
-      current_version: {
-        ...fakeAddon.current_version,
-        compatibility: {},
-      },
-    });
-    const { maxVersion, minVersion } = getCompatibleVersions({
-      addon, clientApp: CLIENT_APP_FIREFOX });
-
-    expect(maxVersion).toEqual(null);
-    expect(minVersion).toEqual(null);
-  });
-
-  it('returns null if current_version does not exist', () => {
-    const addon = createInternalAddon({
-      ...fakeAddon,
-      current_version: null,
-    });
-    const { maxVersion, minVersion } = getCompatibleVersions({
-      addon, clientApp: CLIENT_APP_FIREFOX });
-
-    expect(maxVersion).toEqual(null);
-    expect(minVersion).toEqual(null);
-  });
-
-  it('returns null if addon is null', () => {
-    const { maxVersion, minVersion } = getCompatibleVersions({
-      addon: null, clientApp: CLIENT_APP_FIREFOX });
-
-    expect(maxVersion).toEqual(null);
-    expect(minVersion).toEqual(null);
-  });
-
-  it('should log info when OpenSearch type is found', () => {
-    const fakeLog = { info: sinon.stub() };
-    const openSearchAddon = createInternalAddon({
-      ...fakeAddon,
-      current_version: {
-        ...fakeAddon.current_version,
-        compatibility: {},
-      },
-      type: ADDON_TYPE_OPENSEARCH,
-    });
-    const { maxVersion, minVersion } = getCompatibleVersions({
-      _log: fakeLog,
-      addon: openSearchAddon,
-      clientApp: CLIENT_APP_FIREFOX,
-    });
-
-    expect(maxVersion).toEqual(null);
-    expect(minVersion).toEqual(null);
-    expect(fakeLog.info.firstCall.args[0])
-      .toContain(`addon is type ${ADDON_TYPE_OPENSEARCH}`);
-  });
-});
-
-describe('getClientCompatibility', () => {
-  it('returns true for Firefox (reason undefined when compatibile)', () => {
-    const { browser, os } = UAParser(userAgents.firefox[0]);
-    const userAgentInfo = { browser, os };
-
-    expect(getClientCompatibility({
-      addon: createInternalAddon(fakeAddon),
-      clientApp: CLIENT_APP_FIREFOX,
-      userAgentInfo,
-    })).toEqual({
-      compatible: true,
-      maxVersion: null,
-      minVersion: null,
-      reason: null,
-    });
-  });
-
-  it('returns maxVersion when set', () => {
-    const { browser, os } = UAParser(userAgents.firefox[0]);
-    const userAgentInfo = { browser, os };
-
-    expect(getClientCompatibility({
-      addon: createInternalAddon({
-        ...fakeAddon,
-        current_version: {
-          ...fakeAddon.current_version,
-          compatibility: {
-            firefox: { max: '200.0', min: null },
-          },
-        },
-      }),
-      clientApp: CLIENT_APP_FIREFOX,
-      userAgentInfo,
-    })).toEqual({
-      compatible: true,
-      maxVersion: '200.0',
-      minVersion: null,
-      reason: null,
-    });
-  });
-
-  it('returns minVersion when set', () => {
-    const { browser, os } = UAParser(userAgents.firefox[0]);
-    const userAgentInfo = { browser, os };
-
-    expect(getClientCompatibility({
-      addon: createInternalAddon({
-        ...fakeAddon,
-        current_version: {
-          ...fakeAddon.current_version,
-          compatibility: {
-            firefox: { max: null, min: '2.0' },
-          },
-        },
-      }),
-      clientApp: CLIENT_APP_FIREFOX,
-      userAgentInfo,
-    })).toEqual({
-      compatible: true,
-      maxVersion: null,
-      minVersion: '2.0',
-      reason: null,
-    });
-  });
-
-  it('returns incompatible for non-Firefox UA', () => {
-    const { browser, os } = UAParser(userAgents.firefox[0]);
-    const userAgentInfo = { browser, os };
-
-    expect(getClientCompatibility({
-      addon: createInternalAddon(fakeAddon),
-      clientApp: CLIENT_APP_FIREFOX,
-      userAgentInfo,
-    })).toEqual({
-      compatible: true,
-      maxVersion: null,
-      minVersion: null,
-      reason: null,
-    });
-  });
-
-  it('returns compatible if strict compatibility is off', () => {
-    const { browser, os } = UAParser(userAgents.firefox[4]);
-    const userAgentInfo = { browser, os };
-
-    expect(getClientCompatibility({
-      addon: createInternalAddon({
-        ...fakeAddon,
-        current_version: {
-          ...fakeAddon.current_version,
-          compatibility: {
-            ...fakeAddon.current_version.compatibility,
-            [CLIENT_APP_FIREFOX]: {
-              max: '56.*',
-              min: '24.0',
-            },
-          },
-          files: [{
-            ...fakeAddon.current_version.files[0],
-            is_webextension: true,
-          }],
-          is_strict_compatibility_enabled: false,
-        },
-      }),
-      clientApp: CLIENT_APP_FIREFOX,
-      userAgentInfo,
-    })).toMatchObject({ compatible: true });
-  });
-
-  it('returns incompatible if strict compatibility enabled', () => {
-    const { browser, os } = UAParser(userAgents.firefox[5]);
-    const userAgentInfo = { browser, os };
-
-    expect(getClientCompatibility({
-      addon: createInternalAddon({
-        ...fakeAddon,
-        current_version: {
-          ...fakeAddon.current_version,
-          compatibility: {
-            ...fakeAddon.current_version.compatibility,
-            [CLIENT_APP_FIREFOX]: {
-              max: '56.*',
-              min: '24.0',
-            },
-          },
-          files: [{
-            ...fakeAddon.current_version.files[0],
-            is_webextension: false,
-          }],
-          is_strict_compatibility_enabled: true,
-        },
-      }),
-      clientApp: CLIENT_APP_FIREFOX,
-      userAgentInfo,
-    })).toMatchObject({
-      compatible: false,
-      reason: INCOMPATIBLE_OVER_MAX_VERSION,
-    });
   });
 });
